requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
msgpack==1.0.4
packaging==22.0
Pillow==9.2.0
Ponytile==0.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytz==2022.6
redis==4.3.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
wrapt==1.14.1
zipp==3.11.0
zope.interface==5.5.2
